requirements.txt
# requirements

sqlalchemy==2.0.23
tabulate==0.9.0
twilio==8.11.0
python-dotenv==1.0.1
